Add NSLondon event at AWS
This event is not strictly about WWDC as the talks are focused on server-side Swift. We had the opportunity to put on this event and it seemed like good timing since if a lot of Swift developers get together during WWDC week they‘re going to have a lot of chat about.

I would understand if you think this event isn‘t a good fit.