Add WWDC Notes community project to worldwide events
Last year a separate section was created for "Summaries", but it ended up only listing the WWDC Notes project. But this project really kicks off when the first sessions are out, which usually happens on the second day of WWDC, so I've just put it to June 11th where it really starts.