Add new community projects section for WWDC Notes
I couldn't find any other section this fits in. Last year, we had a "Events around the world and online" section, but I feel this is more of a community learning resource and there might be potentially others (in the future), so I chose this name. Feel free to adjust or move around if needed.